Start working on the planning Engine
The Engine can already calculate if we can deliver the requests, what do
we need to order in order to be able to produce them and what are the
due dates for the products that we can produce given our stock and the
orders.

Also, add a test backend which saves the information obtained from
trello to make testing much faster. It should allow us to make unittests
also.